data-science-project/requirements.txt
# Dependencias del proyecto de detección de neumonía
# Universidad Autónoma de Occidente - Módulo 1

# Core ML libraries
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1

# Image processing
opencv-python==4.8.0.76
Pillow==10.0.0
scikit-image==0.21.0
pydicom==2.3.1

# Data analysis & visualization
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3

# Web framework (para interfaz opcional)
flask==2.3.2
flask-cors==4.0.0

# Utilities
click==8.1.6
python-dotenv==1.0.0
tqdm==4.65.0

# Development & testing
pytest==7.4.0
pytest-cov==4.1.0
typing_extensions>=4.45.0

# Optional: For Jupyter notebooks
jupyter==1.0.0
ipykernel==6.25.0

# System utilities
psutil==5.9.5
